Restructured code and now requires >= Go v1.13 for error handling
In order allow customizations of the golden library per unit test,
the goldie library was restructured to not use global variables
and adopt the function options pattern:

  https://dave.cheney.net/2014/10/17/functional-options-for-friendly-apis

This restructuring also made it easier to implement the following
features:

* Different fixture directories
* A separate directory per unit test
* A separate directory per subtest in the unit test
* Erroring out if a template didn't get data that it needed
* Diff output (classic or in color)
* Using your own diff function for output

Although a little non standard, the methods that a consumer will
be using are now all in the `interface.go` file.

In addition, error handling was cleaned up (now uses pointer
receivers) and leverages the new Go v1.13 `errors` package.

More unit testing is desired.