Don't delete nested directories
*Problem*

Take test cases stuctured in the following way

TestSomeFunctionality
- SubTestA
	- SubSubTestX
	- SubSubTestY
- SubTestB

When SubTestB is updated, it will delete the `TestSomeFunctionality/`
directory. That means the next time we run SubTestA it will have no
Goldie files.

*Solution*
Only delete the files in the current test folder, not the nested
subdirectories. This preserves nested test sub-cases.